major plotting overhaul (first steps)
