+++ libs/tools/average_true_range.py
@@ -2,7 +2,7 @@
 import os
 import pandas as pd
 
-from libs.utils import dual_plotting, INDEXES
+from libs.utils import generate_plot, PlotType, INDEXES
 from libs.features import normalize_signals
 from .moving_average import exponential_moving_avg
 
@@ -85,12 +85,19 @@ def get_atr_signal(fund: pd.DataFrame, **kwargs) -> list:
     if not out_suppress:
         name2 = INDEXES.get(name, name)
         title = f"{name2} - Average True Range"
-        if plot_output:
-            dual_plotting(fund['Close'], signal, 'Price', 'ATR', title=title)
-        else:
-            filename = os.path.join(name, views, f"atr_{name}.png")
-            dual_plotting(fund['Close'], signal, 'Price', 'ATR',
-                          title=title, save_fig=True, filename=filename)
+
+        generate_plot(
+            PlotType.DUAL_PLOTTING,
+            fund['Close'],
+            **dict(
+                y_list_2=signal,
+                y1_label='Price',
+                y2_label='ATR',
+                title=title,
+                filename=os.path.join(name, views, f"atr_{name}.png"),
+                plot_output=plot_output
+            )
+        )
 
     return signal
 
@@ -194,7 +201,14 @@ def atr_indicators(fund: pd.DataFrame, atr_dict: dict, **kwargs) -> dict:
     if plot_output and not out_suppress:
         name2 = INDEXES.get(name, name)
         title = f"{name2} - Average True Range Moving Averages"
-        dual_plotting(fund['Close'], [atr_dict['tabular'],
-                                      ema_1, ema_2], 'Price', 'ATRs', title=title)
+        generate_plot(
+            PlotType.DUAL_PLOTTING,
+            fund['Close'],
+            **dict(
+                y_list_2=[atr_dict['tabular'], ema_1, ema_2],
+                y1_label='Price', y2_label='ATRs', title=title,
+                plot_output=plot_output
+            )
+        )
 
     return atr_dict
